requirements.txt
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2==2.9.9
python-decouple==3.8
Pillow==10.0.1
djangorestframework-gis==1.0
